test: migrate checkpoint.rs unit tests to integration tests
Migrates all 31 unit tests from src/commands/checkpoint.rs to a new
integration test file tests/integration/checkpoint_unit.rs. This is the
largest migration task, handling ~1270 lines of tests.

Changes:
- Made public: FileLineStats, BaseOverrideResolutionPolicy,
  cleanup_failed_captured_checkpoint_prepare, is_ai_author_id,
  compute_file_line_stats, run_with_base_commit_override_with_policy
- Created setup_repo_with_base_commit() helper equivalent to
  TmpRepo::new_with_base_commit()
- Converted all TmpRepo operations to TestRepo + git commands
- All 30 tests passing
