+++ src/commands/checkpoint.rs
@@ -33,11 +33,11 @@ use unicode_normalization::UnicodeNormalization;
 
 /// Per-file line statistics (in-memory only, not persisted)
 #[derive(Debug, Clone, Default)]
-struct FileLineStats {
-    additions: u32,
-    deletions: u32,
-    additions_sloc: u32,
-    deletions_sloc: u32,
+pub struct FileLineStats {
+    pub additions: u32,
+    pub deletions: u32,
+    pub additions_sloc: u32,
+    pub deletions_sloc: u32,
 }
 
 /// Latest checkpoint state needed to process a file in the next checkpoint.
@@ -108,7 +108,7 @@ struct ResolvedCheckpointExecution {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum BaseOverrideResolutionPolicy {
+pub enum BaseOverrideResolutionPolicy {
     AllowFallback,
     RequireExplicitSnapshot,
 }
@@ -243,7 +243,7 @@ fn async_checkpoint_manifest_path(capture_id: &str) -> Result<PathBuf, GitAiErro
     Ok(async_checkpoint_capture_dir(capture_id)?.join("manifest.json"))
 }
 
-fn cleanup_failed_captured_checkpoint_prepare(
+pub fn cleanup_failed_captured_checkpoint_prepare(
     capture_dir: &std::path::Path,
     capture_id: &str,
     error: &GitAiError,
@@ -351,7 +351,7 @@ pub fn run_with_base_commit_override(
 }
 
 #[allow(clippy::too_many_arguments)]
-pub(crate) fn run_with_base_commit_override_with_policy(
+pub fn run_with_base_commit_override_with_policy(
     repo: &Repository,
     author: &str,
     kind: CheckpointKind,
@@ -1579,7 +1579,7 @@ fn content_eq_normalized(a: &str, b: &str) -> bool {
     normalize_line_endings(a) == normalize_line_endings(b)
 }
 
-fn is_ai_author_id(author_id: &str) -> bool {
+pub fn is_ai_author_id(author_id: &str) -> bool {
     author_id != "human" && !author_id.starts_with("h_")
 }
 
@@ -2147,7 +2147,7 @@ fn make_entry_for_file(
 }
 
 /// Compute line statistics for a single file by diffing previous and current content
-fn compute_file_line_stats(previous_content: &str, current_content: &str) -> FileLineStats {
+pub fn compute_file_line_stats(previous_content: &str, current_content: &str) -> FileLineStats {
     let mut stats = FileLineStats::default();
 
     // Use imara_diff to count line changes (matches git's diff algorithm)
@@ -2257,1278 +2257,3 @@ fn is_text_file_in_head(repo: &Repository, path: &str) -> bool {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::authorship::working_log::AgentId;
-    use crate::commands::checkpoint_agent::orchestrator::CheckpointRequest;
-    use crate::git::test_utils::TmpRepo;
-    use std::collections::HashMap;
-    use std::path::PathBuf;
-
-    fn test_checkpoint_request(
-        checkpoint_kind: CheckpointKind,
-        file_paths: Vec<&str>,
-        path_role: PreparedPathRole,
-        dirty_files: Option<HashMap<&str, &str>>,
-    ) -> CheckpointRequest {
-        let agent_id = if checkpoint_kind.is_ai() {
-            Some(AgentId {
-                tool: "test-agent".to_string(),
-                id: "test-capture".to_string(),
-                model: "test-model".to_string(),
-            })
-        } else {
-            None
-        };
-        CheckpointRequest {
-            trace_id: crate::authorship::authorship_log_serialization::generate_trace_id(),
-            checkpoint_kind,
-            agent_id,
-            repo_working_dir: PathBuf::new(),
-            file_paths: file_paths.into_iter().map(PathBuf::from).collect(),
-            path_role,
-            dirty_files: dirty_files.map(|files| {
-                files
-                    .into_iter()
-                    .map(|(path, content)| (PathBuf::from(path), content.to_string()))
-                    .collect()
-            }),
-            transcript_source: None,
-            metadata: HashMap::new(),
-            captured_checkpoint_id: None,
-        }
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_accepts_non_empty_edited_filepaths() {
-        let cr = test_checkpoint_request(
-            CheckpointKind::AiAgent,
-            vec!["src/main.rs"],
-            PreparedPathRole::Edited,
-            None,
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::AiAgent, Some(&cr)),
-            Some((PreparedPathRole::Edited, vec!["src/main.rs".to_string()]))
-        );
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_accepts_non_empty_will_edit_filepaths() {
-        let cr = test_checkpoint_request(
-            CheckpointKind::Human,
-            vec!["src/lib.rs"],
-            PreparedPathRole::WillEdit,
-            None,
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::Human, Some(&cr)),
-            Some((PreparedPathRole::WillEdit, vec!["src/lib.rs".to_string()]))
-        );
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_rejects_dirty_files_without_explicit_paths() {
-        let cr = test_checkpoint_request(
-            CheckpointKind::AiAgent,
-            vec![],
-            PreparedPathRole::Edited,
-            Some(HashMap::from([("src/main.rs", "fn main() {}\n")])),
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::AiAgent, Some(&cr)),
-            None
-        );
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_known_human_uses_edited_filepaths() {
-        // KnownHuman post-save: edit already happened, uses edited (path_role).
-        let cr = test_checkpoint_request(
-            CheckpointKind::KnownHuman,
-            vec!["src/foo.rs"],
-            PreparedPathRole::Edited,
-            None,
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::KnownHuman, Some(&cr)),
-            Some((PreparedPathRole::Edited, vec!["src/foo.rs".to_string()]))
-        );
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_known_human_uses_will_edit_filepaths() {
-        // KnownHuman pre-save: edit hasn't happened yet, uses will_edit (path_role).
-        let cr = test_checkpoint_request(
-            CheckpointKind::KnownHuman,
-            vec!["src/foo.rs"],
-            PreparedPathRole::WillEdit,
-            None,
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::KnownHuman, Some(&cr)),
-            Some((PreparedPathRole::WillEdit, vec!["src/foo.rs".to_string()]))
-        );
-    }
-
-    #[test]
-    fn test_explicit_capture_target_paths_rejects_empty_explicit_lists() {
-        let human_result = test_checkpoint_request(
-            CheckpointKind::Human,
-            vec!["", "   "],
-            PreparedPathRole::WillEdit,
-            None,
-        );
-        let ai_result = test_checkpoint_request(
-            CheckpointKind::AiAgent,
-            vec!["", "   "],
-            PreparedPathRole::Edited,
-            None,
-        );
-
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::Human, Some(&human_result)),
-            None
-        );
-        assert_eq!(
-            explicit_capture_target_paths(CheckpointKind::AiAgent, Some(&ai_result)),
-            None
-        );
-    }
-
-    #[test]
-    fn test_cleanup_failed_captured_checkpoint_prepare_removes_partial_capture_dir() {
-        let temp = tempfile::tempdir().expect("temp dir should be creatable");
-        let capture_dir = temp.path().join("capture-fixture");
-        std::fs::create_dir_all(capture_dir.join("blobs"))
-            .expect("partial capture directory should be creatable");
-        std::fs::write(capture_dir.join("blobs").join("partial-blob"), "partial")
-            .expect("partial blob should be creatable");
-
-        cleanup_failed_captured_checkpoint_prepare(
-            &capture_dir,
-            "capture-fixture",
-            &GitAiError::Generic("synthetic prepare failure".to_string()),
-        );
-
-        assert!(
-            !capture_dir.exists(),
-            "cleanup helper should remove partial capture directories"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_with_staged_changes() {
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Make changes to the file
-        file.append("New line added by user\n").unwrap();
-
-        // Note: TmpFile.append() automatically stages changes (see write_to_disk in test_utils)
-        // So at this point, the file has staged changes
-
-        // Run checkpoint - it should track the changes even though they're staged
-        let (entries_len, files_len, _checkpoints_len) =
-            tmp_repo.trigger_checkpoint_with_author("Aidan").unwrap();
-
-        // The bug: when changes are staged, entries_len is 0 instead of 1
-        assert_eq!(files_len, 1, "Should have 1 file with changes");
-        assert_eq!(
-            entries_len, 1,
-            "Should have 1 file entry in checkpoint (staged changes should be tracked)"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_with_staged_changes_after_previous_checkpoint() {
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Make first changes and checkpoint
-        file.append("First change\n").unwrap();
-        let (entries_len_1, files_len_1, _) =
-            tmp_repo.trigger_checkpoint_with_author("Aidan").unwrap();
-
-        assert_eq!(
-            files_len_1, 1,
-            "First checkpoint: should have 1 file with changes"
-        );
-        assert_eq!(
-            entries_len_1, 1,
-            "First checkpoint: should have 1 file entry"
-        );
-
-        // Make second changes - these are already staged by append()
-        file.append("Second change\n").unwrap();
-
-        // Run checkpoint again - it should track the staged changes even after a previous checkpoint
-        let (entries_len_2, files_len_2, _) =
-            tmp_repo.trigger_checkpoint_with_author("Aidan").unwrap();
-
-        assert_eq!(
-            files_len_2, 1,
-            "Second checkpoint: should have 1 file with changes"
-        );
-        assert_eq!(
-            entries_len_2, 1,
-            "Second checkpoint: should have 1 file entry in checkpoint (staged changes should be tracked)"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_with_only_staged_no_unstaged_changes() {
-        use std::fs;
-
-        // Create a repo with an initial commit
-        let (tmp_repo, file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Get the file path
-        let file_path = file.path();
-        let filename = file.filename();
-
-        // Manually modify the file (bypassing TmpFile's automatic staging)
-        let mut content = fs::read_to_string(&file_path).unwrap();
-        content.push_str("New line for staging test\n");
-        fs::write(&file_path, &content).unwrap();
-
-        // Now manually stage it using git (this is what "git add" does)
-        tmp_repo.stage_file(filename).unwrap();
-
-        // At this point: HEAD has old content, index has new content, workdir has new content
-        // And unstaged should be "Unmodified" because workdir == index
-
-        // Now run checkpoint
-        let (entries_len, files_len, _checkpoints_len) =
-            tmp_repo.trigger_checkpoint_with_author("Aidan").unwrap();
-
-        // This should work: we should see 1 file with 1 entry
-        assert_eq!(files_len, 1, "Should detect 1 file with staged changes");
-        assert_eq!(
-            entries_len, 1,
-            "Should track the staged changes in checkpoint"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_with_only_unstaged_changes_for_ai_without_pathspec() {
-        use std::fs;
-
-        // Create a repo with an initial commit
-        let (tmp_repo, file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Manually modify the file without staging it
-        let file_path = file.path();
-        let mut content = fs::read_to_string(&file_path).unwrap();
-        content.push_str("New unstaged AI line\n");
-        fs::write(&file_path, &content).unwrap();
-
-        // Trigger AI checkpoint without edited_filepaths (pathspec-less flow used by some agents)
-        let (entries_len, files_len, _checkpoints_len) = tmp_repo
-            .trigger_checkpoint_with_ai("Codex", Some("gpt-5-codex"), Some("codex"))
-            .unwrap();
-
-        assert_eq!(
-            files_len, 1,
-            "Should detect unstaged changes without pathspecs"
-        );
-        assert_eq!(
-            entries_len, 1,
-            "Should create an AI checkpoint entry for unstaged changes without pathspecs"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_base_override_controls_head_context_for_entry_generation() {
-        use crate::authorship::working_log::AgentId;
-        use crate::commands::checkpoint_agent::orchestrator::CheckpointRequest;
-        use std::collections::HashMap;
-        use std::fs;
-
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-        let filename = file.filename().to_string();
-
-        file.update("line from commit A\n").unwrap();
-        tmp_repo.commit_with_message("commit A").unwrap();
-        let base_commit = tmp_repo.get_head_commit_sha().unwrap();
-
-        file.update("line from commit B\n").unwrap();
-        tmp_repo.commit_with_message("commit B").unwrap();
-
-        // Keep the worktree dirty so git status returns this file, but inject deterministic
-        // content from commit B via dirty_files.
-        fs::write(file.path(), "line from uncommitted edit\n").unwrap();
-
-        let mut dirty_files = HashMap::new();
-        dirty_files.insert(
-            PathBuf::from(filename.clone()),
-            "line from commit B\n".to_string(),
-        );
-        let cr = CheckpointRequest {
-            trace_id: crate::authorship::authorship_log_serialization::generate_trace_id(),
-            checkpoint_kind: CheckpointKind::AiAgent,
-            agent_id: Some(AgentId {
-                tool: "mock_ai".to_string(),
-                id: "base-override-regression".to_string(),
-                model: "test".to_string(),
-            }),
-            repo_working_dir: PathBuf::new(),
-            file_paths: vec![PathBuf::from(filename)],
-            path_role: PreparedPathRole::Edited,
-            dirty_files: Some(dirty_files),
-            transcript_source: None,
-            metadata: HashMap::new(),
-            captured_checkpoint_id: None,
-        };
-
-        let (entries_len, files_len, _) = run_with_base_commit_override_with_policy(
-            tmp_repo.gitai_repo(),
-            "mock-ai",
-            CheckpointKind::AiAgent,
-            true,
-            Some(cr),
-            false,
-            Some(base_commit.as_str()),
-            BaseOverrideResolutionPolicy::RequireExplicitSnapshot,
-        )
-        .unwrap();
-
-        assert_eq!(
-            files_len, 1,
-            "Expected one tracked file for the checkpoint run"
-        );
-        assert_eq!(
-            entries_len, 1,
-            "When base override points to commit A, current content from commit B must produce an entry"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_base_override_strict_rejects_missing_dirty_snapshot() {
-        use crate::authorship::working_log::AgentId;
-        use crate::commands::checkpoint_agent::orchestrator::CheckpointRequest;
-        use std::fs;
-
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-        let filename = file.filename().to_string();
-
-        file.update("line from commit A\n").unwrap();
-        tmp_repo.commit_with_message("commit A").unwrap();
-        let base_commit = tmp_repo.get_head_commit_sha().unwrap();
-
-        file.update("line from commit B\n").unwrap();
-        tmp_repo.commit_with_message("commit B").unwrap();
-
-        // Keep the worktree dirty so the legacy fallback would succeed if it were used.
-        fs::write(file.path(), "line from uncommitted edit\n").unwrap();
-
-        let cr = CheckpointRequest {
-            trace_id: crate::authorship::authorship_log_serialization::generate_trace_id(),
-            checkpoint_kind: CheckpointKind::AiAgent,
-            agent_id: Some(AgentId {
-                tool: "mock_ai".to_string(),
-                id: "base-override-strict-missing-snapshot".to_string(),
-                model: "test".to_string(),
-            }),
-            repo_working_dir: PathBuf::new(),
-            file_paths: vec![PathBuf::from(filename.clone())],
-            path_role: PreparedPathRole::Edited,
-            dirty_files: None,
-            transcript_source: None,
-            metadata: HashMap::new(),
-            captured_checkpoint_id: None,
-        };
-
-        let error = run_with_base_commit_override_with_policy(
-            tmp_repo.gitai_repo(),
-            "mock-ai",
-            CheckpointKind::AiAgent,
-            true,
-            Some(cr),
-            false,
-            Some(base_commit.as_str()),
-            BaseOverrideResolutionPolicy::RequireExplicitSnapshot,
-        )
-        .expect_err("strict base override should reject missing dirty snapshots");
-
-        assert!(
-            error.to_string().contains(
-                "requires explicit in-repository target paths and a matching dirty snapshot"
-            ),
-            "expected strict snapshot error, got: {}",
-            error
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_base_override_allow_fallback_scans_when_snapshot_missing() {
-        use crate::authorship::working_log::AgentId;
-        use crate::commands::checkpoint_agent::orchestrator::CheckpointRequest;
-        use std::fs;
-
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-        let filename = file.filename().to_string();
-
-        file.update("line from commit A\n").unwrap();
-        tmp_repo.commit_with_message("commit A").unwrap();
-        let base_commit = tmp_repo.get_head_commit_sha().unwrap();
-
-        file.update("line from commit B\n").unwrap();
-        tmp_repo.commit_with_message("commit B").unwrap();
-
-        // Without a dirty snapshot the fallback path must rediscover the dirty file from the repo.
-        fs::write(file.path(), "line from uncommitted edit\n").unwrap();
-
-        let cr = CheckpointRequest {
-            trace_id: crate::authorship::authorship_log_serialization::generate_trace_id(),
-            checkpoint_kind: CheckpointKind::AiAgent,
-            agent_id: Some(AgentId {
-                tool: "mock_ai".to_string(),
-                id: "base-override-allow-fallback".to_string(),
-                model: "test".to_string(),
-            }),
-            repo_working_dir: PathBuf::new(),
-            file_paths: vec![PathBuf::from(filename)],
-            path_role: PreparedPathRole::Edited,
-            dirty_files: None,
-            transcript_source: None,
-            metadata: HashMap::new(),
-            captured_checkpoint_id: None,
-        };
-
-        let (entries_len, files_len, _) = run_with_base_commit_override(
-            tmp_repo.gitai_repo(),
-            "mock-ai",
-            CheckpointKind::AiAgent,
-            true,
-            Some(cr),
-            false,
-            Some(base_commit.as_str()),
-        )
-        .expect("allow-fallback base override should still scan the repo");
-
-        assert_eq!(
-            files_len, 1,
-            "fallback path should rediscover the changed file"
-        );
-        assert_eq!(
-            entries_len, 1,
-            "fallback path should still produce checkpoint entries from the worktree scan"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_skips_conflicted_files() {
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Get the current branch name (whatever the default is)
-        let base_branch = tmp_repo.current_branch().unwrap();
-
-        // Create a branch and make different changes on each branch to create a conflict
-        tmp_repo.create_branch("feature-branch").unwrap();
-
-        // On feature branch, modify the file
-        file.append("Feature branch change\n").unwrap();
-        tmp_repo
-            .trigger_checkpoint_with_author("FeatureUser")
-            .unwrap();
-        tmp_repo.commit_with_message("Feature commit").unwrap();
-
-        // Switch back to base branch and make conflicting changes
-        tmp_repo.switch_branch(&base_branch).unwrap();
-        file.append("Main branch change\n").unwrap();
-        tmp_repo.trigger_checkpoint_with_author("MainUser").unwrap();
-        tmp_repo.commit_with_message("Main commit").unwrap();
-
-        // Attempt to merge feature-branch into base branch - this should create a conflict
-        let has_conflicts = tmp_repo.merge_with_conflicts("feature-branch").unwrap();
-        assert!(has_conflicts, "Should have merge conflicts");
-
-        // Try to checkpoint while there are conflicts
-        let (entries_len, files_len, _) = tmp_repo.trigger_checkpoint_with_author("Human").unwrap();
-
-        // Checkpoint should skip conflicted files
-        assert_eq!(
-            files_len, 0,
-            "Should have 0 files (conflicted file should be skipped)"
-        );
-        assert_eq!(
-            entries_len, 0,
-            "Should have 0 entries (conflicted file should be skipped)"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_with_paths_outside_repo() {
-        use crate::authorship::working_log::AgentId;
-        use crate::commands::checkpoint_agent::orchestrator::CheckpointRequest;
-
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Make changes to the file
-        file.append("New line added\n").unwrap();
-
-        // Create checkpoint result with paths outside the repo
-        let cr = CheckpointRequest {
-            trace_id: crate::authorship::authorship_log_serialization::generate_trace_id(),
-            checkpoint_kind: CheckpointKind::AiAgent,
-            agent_id: Some(AgentId {
-                tool: "test_tool".to_string(),
-                id: "test_session".to_string(),
-                model: "test_model".to_string(),
-            }),
-            repo_working_dir: PathBuf::new(),
-            file_paths: vec![
-                PathBuf::from("/tmp/outside_file.txt"),
-                PathBuf::from("../outside_parent.txt"),
-                PathBuf::from(file.filename()), // This one is valid
-            ],
-            path_role: PreparedPathRole::Edited,
-            dirty_files: None,
-            transcript_source: None,
-            metadata: HashMap::new(),
-            captured_checkpoint_id: None,
-        };
-
-        // Run checkpoint - should not crash even with paths outside repo
-        let result = tmp_repo.trigger_checkpoint_with_checkpoint_request("test_user", Some(cr));
-
-        // Should succeed without crashing
-        assert!(
-            result.is_ok(),
-            "Checkpoint should succeed even with paths outside repo: {:?}",
-            result.err()
-        );
-
-        let (entries_len, files_len, _) = result.unwrap();
-        // Should only process the valid file
-        assert_eq!(files_len, 1, "Should process 1 valid file");
-        assert_eq!(entries_len, 1, "Should create 1 entry");
-    }
-
-    #[test]
-    fn test_checkpoint_filters_external_paths_from_stored_checkpoints() {
-        use crate::authorship::working_log::{Checkpoint, WorkingLogEntry};
-
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Get access to the working log storage
-        let repo =
-            crate::git::repository::find_repository_in_path(tmp_repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-
-        // Manually inject a checkpoint with an external file path (simulating the bug)
-        // This is what happens when a file outside the repo was tracked before the fix
-        let working_log = repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-
-        let external_entry = WorkingLogEntry::new(
-            "/external/path/outside/repo.txt".to_string(),
-            "fake_sha_for_external".to_string(),
-            vec![],
-            vec![],
-        );
-
-        let fake_checkpoint = Checkpoint::new(
-            CheckpointKind::Human,
-            "fake_diff".to_string(),
-            "test_author".to_string(),
-            vec![external_entry],
-        );
-
-        // Store the checkpoint with external path
-        working_log
-            .append_checkpoint(&fake_checkpoint)
-            .expect("Should be able to append checkpoint");
-
-        // Now make actual changes to a file in the repo
-        file.append("New line for testing\n").unwrap();
-
-        // Run checkpoint - this should NOT crash even though there's an external path stored
-        // Previously this would fail with: "fatal: /external/path/outside/repo.txt is outside repository"
-        let result = tmp_repo.trigger_checkpoint_with_author("Human");
-
-        assert!(
-            result.is_ok(),
-            "Checkpoint should succeed even with external paths stored in previous checkpoints: {:?}",
-            result.err()
-        );
-
-        let (entries_len, files_len, _) = result.unwrap();
-        // Should only process the valid file in the repo
-        assert_eq!(
-            files_len, 1,
-            "Should process 1 valid file (external path should be filtered)"
-        );
-        assert_eq!(entries_len, 1, "Should create 1 entry for the in-repo file");
-    }
-
-    #[test]
-    fn test_checkpoint_works_after_conflict_resolution_maintains_authorship() {
-        // Create a repo with an initial commit
-        let (tmp_repo, mut file, _) = TmpRepo::new_with_base_commit().unwrap();
-
-        // Get the current branch name (whatever the default is)
-        let base_branch = tmp_repo.current_branch().unwrap();
-
-        // Checkpoint initial state to track the base authorship
-        let file_path = file.path();
-        let initial_content = std::fs::read_to_string(&file_path).unwrap();
-        println!("Initial content:\n{}", initial_content);
-
-        // Create a branch and make changes
-        tmp_repo.create_branch("feature-branch").unwrap();
-        file.append("Feature line 1\n").unwrap();
-        file.append("Feature line 2\n").unwrap();
-        tmp_repo.trigger_checkpoint_with_author("AI_Agent").unwrap();
-        tmp_repo.commit_with_message("Feature commit").unwrap();
-
-        // Switch back to base branch and make conflicting changes
-        tmp_repo.switch_branch(&base_branch).unwrap();
-        file.append("Main line 1\n").unwrap();
-        file.append("Main line 2\n").unwrap();
-        tmp_repo.trigger_checkpoint_with_author("Human").unwrap();
-        tmp_repo.commit_with_message("Main commit").unwrap();
-
-        // Attempt to merge feature-branch into base branch - this should create a conflict
-        let has_conflicts = tmp_repo.merge_with_conflicts("feature-branch").unwrap();
-        assert!(has_conflicts, "Should have merge conflicts");
-
-        // While there are conflicts, checkpoint should skip the file
-        let (entries_len_conflict, files_len_conflict, _) =
-            tmp_repo.trigger_checkpoint_with_author("Human").unwrap();
-        assert_eq!(
-            files_len_conflict, 0,
-            "Should skip conflicted files during conflict"
-        );
-        assert_eq!(
-            entries_len_conflict, 0,
-            "Should not create entries for conflicted files"
-        );
-
-        // Resolve the conflict by choosing "ours" (base branch)
-        tmp_repo.resolve_conflict(file.filename(), "ours").unwrap();
-
-        // Verify content to ensure the resolution was applied correctly
-        let resolved_content = std::fs::read_to_string(&file_path).unwrap();
-        println!("Resolved content after resolution:\n{}", resolved_content);
-        assert!(
-            resolved_content.contains("Main line 1"),
-            "Should contain base branch content (we chose 'ours')"
-        );
-        assert!(
-            resolved_content.contains("Main line 2"),
-            "Should contain base branch content (we chose 'ours')"
-        );
-        assert!(
-            !resolved_content.contains("Feature line 1"),
-            "Should not contain feature branch content (we chose 'ours')"
-        );
-
-        // After resolution, make additional changes to test that checkpointing works again
-        file.append("Post-resolution line 1\n").unwrap();
-        file.append("Post-resolution line 2\n").unwrap();
-
-        // Now checkpoint should work and track the new changes
-        let (entries_len_after, files_len_after, _) =
-            tmp_repo.trigger_checkpoint_with_author("Human").unwrap();
-
-        println!(
-            "After resolution and new changes: entries_len={}, files_len={}",
-            entries_len_after, files_len_after
-        );
-
-        // The file should be tracked with the new changes
-        assert_eq!(
-            files_len_after, 1,
-            "Should detect 1 file with new changes after conflict resolution"
-        );
-        assert_eq!(
-            entries_len_after, 1,
-            "Should create 1 entry for new changes after conflict resolution"
-        );
-    }
-
-    #[test]
-    fn test_known_human_checkpoint_without_ai_history_records_h_hash_attributions() {
-        let repo = TmpRepo::new().unwrap();
-        let mut file = repo.write_file("simple.txt", "one\n", true).unwrap();
-
-        repo.trigger_checkpoint_with_author("seed").unwrap();
-        repo.commit_with_message("seed commit").unwrap();
-
-        file.append("two\n").unwrap();
-        repo.trigger_checkpoint_with_author("human").unwrap();
-
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-        let latest = checkpoints.last().unwrap();
-        let entry = latest
-            .entries
-            .iter()
-            .find(|entry| entry.file == "simple.txt")
-            .unwrap();
-
-        // KnownHuman checkpoints always record h_<hash> line attributions, even with no AI history.
-        // This allows downstream stats to count these lines as human_additions.
-        assert!(
-            !entry.line_attributions.is_empty(),
-            "KnownHuman checkpoint should record line-level h_<hash> attributions"
-        );
-        assert!(
-            entry
-                .line_attributions
-                .iter()
-                .all(|la| la.author_id.starts_with("h_")),
-            "All line attributions should be h_<hash> IDs"
-        );
-        assert!(
-            latest.line_stats.additions > 0,
-            "KnownHuman checkpoint should record line stats"
-        );
-    }
-
-    #[test]
-    fn test_human_checkpoint_keeps_attributions_for_ai_touched_file() {
-        let (repo, mut lines_file, mut alphabet_file) = TmpRepo::new_with_base_commit().unwrap();
-
-        lines_file.append("ai change\n").unwrap();
-        repo.trigger_checkpoint_with_ai("mock_ai", None, None)
-            .unwrap();
-
-        lines_file.append("human after ai\n").unwrap();
-        alphabet_file.append("human only\n").unwrap();
-        repo.trigger_checkpoint_with_author("human").unwrap();
-
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-        let latest = checkpoints.last().unwrap();
-
-        let ai_touched_entry = latest
-            .entries
-            .iter()
-            .find(|entry| entry.file == "lines.md")
-            .unwrap();
-        assert!(
-            !ai_touched_entry.attributions.is_empty()
-                || !ai_touched_entry.line_attributions.is_empty(),
-            "AI-touched file should keep attribution tracking"
-        );
-
-        let human_only_entry = latest
-            .entries
-            .iter()
-            .find(|entry| entry.file == "alphabet.md")
-            .unwrap();
-        // KnownHuman checkpoints record h_<hash> attributions for all files, including
-        // files with no AI history. This ensures human lines are counted correctly in stats.
-        assert!(
-            !human_only_entry.line_attributions.is_empty(),
-            "KnownHuman checkpoint should record line attributions for human-only files"
-        );
-        assert!(
-            human_only_entry
-                .line_attributions
-                .iter()
-                .all(|la| la.author_id.starts_with("h_")),
-            "Human-only file attributions should all be h_<hash> IDs"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_skips_default_ignored_files() {
-        let repo = TmpRepo::new().unwrap();
-        repo.write_file("README.md", "# repo\n", true).unwrap();
-        repo.commit_with_message("initial").unwrap();
-
-        std::fs::write(repo.path().join("README.md"), "# repo\n\nupdated\n").unwrap();
-        std::fs::write(repo.path().join("Cargo.lock"), "# lock\n# lock2\n").unwrap();
-
-        repo.trigger_checkpoint_with_author("human").unwrap();
-
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-        let latest = checkpoints.last().unwrap();
-
-        assert!(
-            latest.entries.iter().any(|entry| entry.file == "README.md"),
-            "Expected non-ignored source file to be checkpointed"
-        );
-        assert!(
-            latest
-                .entries
-                .iter()
-                .all(|entry| entry.file != "Cargo.lock"),
-            "Expected Cargo.lock to be filtered by default ignore patterns"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_skips_linguist_generated_files_from_root_gitattributes() {
-        let repo = TmpRepo::new().unwrap();
-        repo.write_file("README.md", "# repo\n", true).unwrap();
-        repo.commit_with_message("initial").unwrap();
-
-        repo.write_file(".gitattributes", "generated/** linguist-generated\n", true)
-            .unwrap();
-        repo.commit_with_message("attrs").unwrap();
-
-        std::fs::create_dir_all(repo.path().join("generated")).unwrap();
-        std::fs::write(
-            repo.path().join("generated").join("api.generated.ts"),
-            "// generated\n// generated 2\n",
-        )
-        .unwrap();
-        repo.write_file("main.rs", "fn main() {}\n", false).unwrap();
-
-        repo.trigger_checkpoint_with_author("human").unwrap();
-
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-        let latest = checkpoints.last().unwrap();
-
-        assert!(
-            latest.entries.iter().any(|entry| entry.file == "main.rs"),
-            "Expected non-generated file to be checkpointed"
-        );
-        assert!(
-            latest
-                .entries
-                .iter()
-                .all(|entry| entry.file != "generated/api.generated.ts"),
-            "Expected linguist-generated file to be filtered via .gitattributes"
-        );
-    }
-
-    #[test]
-    fn test_compute_line_stats_ignores_whitespace_only_lines() {
-        let (tmp_repo, _lines_file, _alphabet_file) = TmpRepo::new_with_base_commit().unwrap();
-
-        let repo =
-            crate::git::repository::find_repository_in_path(tmp_repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-
-        let base_commit = repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-
-        let mut test_file = tmp_repo
-            .write_file("whitespace.txt", "Seed line\n", true)
-            .unwrap();
-
-        tmp_repo
-            .trigger_checkpoint_with_author("Setup")
-            .expect("Setup checkpoint should succeed");
-
-        test_file
-            .append("\n\n   \nVisible line one\n\n\t\nVisible line two\n  \n")
-            .unwrap();
-
-        tmp_repo
-            .trigger_checkpoint_with_author("Aidan")
-            .expect("First checkpoint should succeed");
-
-        let after_add_stats = working_log
-            .read_all_checkpoints()
-            .expect("Should read checkpoints after addition");
-        let after_add_last = after_add_stats
-            .last()
-            .expect("At least one checkpoint expected")
-            .line_stats
-            .clone();
-
-        assert_eq!(
-            after_add_last.additions, 8,
-            "Additions includes empty lines"
-        );
-        assert_eq!(after_add_last.deletions, 0, "No deletions expected yet");
-        assert_eq!(
-            after_add_last.additions_sloc, 2,
-            "Only visible lines counted"
-        );
-        assert_eq!(
-            after_add_last.deletions_sloc, 0,
-            "No deletions expected yet"
-        );
-
-        let cleaned_content = std::fs::read_to_string(test_file.path()).unwrap();
-        let cleaned_lines: Vec<&str> = cleaned_content
-            .lines()
-            .filter(|line| !line.trim().is_empty())
-            .collect();
-        let cleaned_body = format!("{}\n", cleaned_lines.join("\n"));
-        test_file.update(&cleaned_body).unwrap();
-
-        tmp_repo
-            .trigger_checkpoint_with_author("Aidan")
-            .expect("Second checkpoint should succeed");
-
-        let after_delete_stats = working_log
-            .read_all_checkpoints()
-            .expect("Should read checkpoints after deletion");
-        let latest_stats = after_delete_stats
-            .last()
-            .expect("At least one checkpoint expected")
-            .line_stats
-            .clone();
-
-        assert_eq!(
-            latest_stats.additions, 0,
-            "No additions in cleanup checkpoint"
-        );
-        assert_eq!(latest_stats.deletions, 6, "Deletions includes empty lines");
-        assert_eq!(
-            latest_stats.additions_sloc, 0,
-            "No additions in cleanup checkpoint"
-        );
-        assert_eq!(
-            latest_stats.deletions_sloc, 0,
-            "Whitespace deletions ignored"
-        );
-    }
-
-    // ====================================================================
-    // CRLF / LF normalization tests for compute_file_line_stats
-    // ====================================================================
-
-    #[test]
-    fn test_compute_file_line_stats_crlf_to_lf_no_changes() {
-        // Same content, only line endings differ (CRLF → LF).
-        // Stats should show 0 additions and 0 deletions.
-        let old = "line1\r\nline2\r\nline3\r\n";
-        let new = "line1\nline2\nline3\n";
-
-        let stats = compute_file_line_stats(old, new);
-
-        assert_eq!(
-            stats.additions, 0,
-            "CRLF→LF with identical content should show 0 additions"
-        );
-        assert_eq!(
-            stats.deletions, 0,
-            "CRLF→LF with identical content should show 0 deletions"
-        );
-    }
-
-    #[test]
-    fn test_compute_file_line_stats_lf_to_crlf_no_changes() {
-        let old = "line1\nline2\nline3\n";
-        let new = "line1\r\nline2\r\nline3\r\n";
-
-        let stats = compute_file_line_stats(old, new);
-
-        assert_eq!(
-            stats.additions, 0,
-            "LF→CRLF with identical content should show 0 additions"
-        );
-        assert_eq!(
-            stats.deletions, 0,
-            "LF→CRLF with identical content should show 0 deletions"
-        );
-    }
-
-    #[test]
-    fn test_compute_file_line_stats_crlf_to_lf_with_additions() {
-        // Reproduces the user-reported bug: file with CRLF, AI adds lines with LF.
-        // Old: 3 CRLF lines. New: same 3 lines (LF) + 2 new lines.
-        // Should show exactly 2 additions and 0 deletions.
-        let old = "line1\r\nline2\r\nline3\r\n";
-        let new = "line1\nline2\nline3\nnew_a\nnew_b\n";
-
-        let stats = compute_file_line_stats(old, new);
-
-        assert_eq!(
-            stats.additions, 2,
-            "Should have exactly 2 additions (the new lines)"
-        );
-        assert_eq!(
-            stats.deletions, 0,
-            "Should have 0 deletions (no lines removed)"
-        );
-    }
-
-    #[test]
-    fn test_compute_file_line_stats_crlf_large_file_user_reported_bug() {
-        // Exact scenario from user report:
-        // 100-line CRLF file, AI adds 5 lines (with LF).
-        // Should show +5 -0, NOT +105 -100.
-        let mut old = String::new();
-        for i in 1..=100 {
-            old.push_str(&format!("line number {}\r\n", i));
-        }
-
-        let mut new = String::new();
-        for i in 1..=100 {
-            new.push_str(&format!("line number {}\n", i));
-        }
-        for i in 1..=5 {
-            new.push_str(&format!("new ai line {}\n", i));
-        }
-
-        let stats = compute_file_line_stats(&old, &new);
-
-        assert_eq!(
-            stats.additions, 5,
-            "Should have exactly 5 additions (AI-added lines), not {}",
-            stats.additions
-        );
-        assert_eq!(
-            stats.deletions, 0,
-            "Should have 0 deletions, not {}",
-            stats.deletions
-        );
-    }
-
-    // ====================================================================
-    // End-to-end CRLF test: blob has CRLF, working tree has LF
-    // Simulates the real-world scenario where git stores CRLF (or autocrlf
-    // converts on checkout) and an AI tool writes LF.
-    // ====================================================================
-
-    #[test]
-    fn test_checkpoint_crlf_blob_vs_lf_working_tree_stats_not_inflated() {
-        // Step 1: Create a repo and commit a file with CRLF line endings.
-        // On Linux without autocrlf, the blob stores CRLF verbatim.
-        let repo = TmpRepo::new().unwrap();
-        let crlf_content = "line1\r\nline2\r\nline3\r\nline4\r\nline5\r\n";
-        repo.write_file("test.txt", crlf_content, true).unwrap();
-        repo.commit_with_message("initial commit with CRLF")
-            .unwrap();
-
-        // Step 2: Overwrite the file with LF endings + one new line,
-        // simulating an AI tool that writes LF on a Windows repo.
-        let lf_content_with_addition = "line1\nline2\nline3\nline4\nline5\nnew_ai_line\n";
-        std::fs::write(repo.path().join("test.txt"), lf_content_with_addition).unwrap();
-
-        // Step 3: Run a checkpoint
-        repo.trigger_checkpoint_with_author("test-author").unwrap();
-
-        // Step 4: Read back checkpoint stats
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-        let latest = checkpoints
-            .last()
-            .expect("Should have at least one checkpoint");
-
-        // The key assertion: stats should reflect only the actual addition,
-        // NOT inflate every line because of CRLF→LF conversion.
-        assert_eq!(
-            latest.line_stats.additions, 1,
-            "Should have 1 addition (the new AI line), not {} (which would mean CRLF→LF inflated the count)",
-            latest.line_stats.additions
-        );
-        assert_eq!(
-            latest.line_stats.deletions, 0,
-            "Should have 0 deletions, not {} (which would mean CRLF→LF caused all old lines to appear deleted)",
-            latest.line_stats.deletions
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_crlf_blob_vs_lf_working_tree_no_changes_skipped() {
-        // When the only difference is CRLF→LF (no actual content change),
-        // the checkpoint should skip the file entirely — content_eq_normalized
-        // detects they're equal and returns None.
-        let repo = TmpRepo::new().unwrap();
-        let crlf_content = "line1\r\nline2\r\nline3\r\n";
-        repo.write_file("test.txt", crlf_content, true).unwrap();
-        repo.commit_with_message("initial commit with CRLF")
-            .unwrap();
-
-        // Overwrite with LF-only — same text content, different line endings
-        let lf_content = "line1\nline2\nline3\n";
-        std::fs::write(repo.path().join("test.txt"), lf_content).unwrap();
-
-        repo.trigger_checkpoint_with_author("test-author").unwrap();
-
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-
-        // The checkpoint may be empty (no entries) or absent entirely,
-        // because content_eq_normalized correctly detected no real change.
-        if let Some(latest) = checkpoints.last() {
-            let test_entry = latest.entries.iter().find(|e| e.file == "test.txt");
-            assert!(
-                test_entry.is_none(),
-                "test.txt should be skipped when only line endings differ"
-            );
-        }
-        // If no checkpoints at all, that's also correct — nothing changed.
-    }
-
-    #[test]
-    fn test_checkpoint_stale_crlf_blob_causes_ai_reattribution() {
-        // Regression test for Devin review finding: when a CRLF-only change is
-        // skipped (preserving a stale CRLF blob), the NEXT AI checkpoint compares
-        // the stale CRLF blob against the LF working tree. Because
-        // capture_diff_slices sees "line\r\n" ≠ "line\n", ALL lines appear changed.
-        // With force_split=true in AI checkpoints, every "changed" line gets
-        // re-attributed to AI — even human-written lines.
-        //
-        // The fix: when content differs only in line endings, update the blob
-        // to LF (preserving attributions) so future diffs are LF-vs-LF.
-        let repo = TmpRepo::new().unwrap();
-        let crlf_initial = "human_line1\r\nhuman_line2\r\nhuman_line3\r\n";
-        repo.write_file("test.txt", crlf_initial, true).unwrap();
-        repo.commit_with_message("initial commit with CRLF")
-            .unwrap();
-
-        // Step 1: Human checkpoint on CRLF file → creates entry with CRLF blob
-        // (need to add a line so the checkpoint creates an entry)
-        let crlf_with_edit = "human_line1\r\nhuman_line2\r\nhuman_line3\r\nhuman_line4\r\n";
-        std::fs::write(repo.path().join("test.txt"), crlf_with_edit).unwrap();
-        repo.trigger_checkpoint_with_author("human-author").unwrap();
-
-        // Step 2: Convert file to LF (same content, only line endings change)
-        let lf_with_edit = "human_line1\nhuman_line2\nhuman_line3\nhuman_line4\n";
-        std::fs::write(repo.path().join("test.txt"), lf_with_edit).unwrap();
-        repo.trigger_checkpoint_with_author("human-author").unwrap();
-
-        // Step 3: AI adds one line (LF) → AI checkpoint
-        let lf_with_ai = "human_line1\nhuman_line2\nhuman_line3\nhuman_line4\nai_new_line\n";
-        std::fs::write(repo.path().join("test.txt"), lf_with_ai).unwrap();
-        repo.trigger_checkpoint_with_ai("Claude", None, None)
-            .unwrap();
-
-        // Read the AI checkpoint
-        let gitai_repo =
-            crate::git::repository::find_repository_in_path(repo.path().to_str().unwrap())
-                .expect("Repository should exist");
-        let base_commit = gitai_repo
-            .head()
-            .ok()
-            .and_then(|head| head.target().ok())
-            .unwrap_or_else(|| "initial".to_string());
-        let working_log = gitai_repo
-            .storage
-            .working_log_for_base_commit(&base_commit)
-            .unwrap();
-        let checkpoints = working_log.read_all_checkpoints().unwrap();
-
-        // Find the AI checkpoint entry for test.txt
-        let ai_checkpoint = checkpoints
-            .iter()
-            .rev()
-            .find(|cp| cp.kind.is_ai() && cp.entries.iter().any(|e| e.file == "test.txt"))
-            .expect("Should have an AI checkpoint with test.txt");
-        let test_entry = ai_checkpoint
-            .entries
-            .iter()
-            .find(|e| e.file == "test.txt")
-            .unwrap();
-
-        // The key assertion: the AI checkpoint should NOT attribute all lines to AI.
-        // Only the actually-added line should be AI-attributed.
-        let ai_line_attrs: Vec<_> = test_entry
-            .line_attributions
-            .iter()
-            .filter(|la| is_ai_author_id(&la.author_id))
-            .collect();
-
-        // Count total lines covered by AI attributions
-        let ai_line_count: u32 = ai_line_attrs
-            .iter()
-            .map(|la| la.end_line - la.start_line + 1)
-            .sum();
-
-        // AI should only attribute 1 line (the new ai_new_line), not all 5 lines.
-        // If the stale CRLF blob caused full re-attribution, ai_line_count would be 5.
-        assert!(
-            ai_line_count <= 2,
-            "AI should attribute at most 1-2 lines (the actual addition), \
-             but attributed {} lines — stale CRLF blob caused full re-attribution. \
-             AI attributions: {:?}, all attributions: {:?}",
-            ai_line_count,
-            ai_line_attrs,
-            test_entry.line_attributions
-        );
-    }
-}
