+++ tests/integration/main.rs
@@ -23,6 +23,7 @@ mod blame_subdirectory;
 mod checkout_switch;
 mod checkpoint_explicit_paths;
 mod checkpoint_perf;
+mod checkpoint_unit;
 mod checkpoint_size;
 mod cherry_pick;
 mod chinese_text_edits;
